♻️ Refactor video block ui
